move check system requirements to workload
